# Support deterministic scheduling