See python-trio/pytest-trio#73 - this is important for Hypothesis, and arguably for debugging in general.  

I've used the current time as primary sort key, with name and task cancel/schedule points to order tasks that started at the same time according to perf_counter.  This is all deterministic, and any remaining collisions are probably coming from random bitflips.

I've also designed the tests to be imported into pytest-trio's test suite, and the Hypothesis test added.