@Zac-HD 2019-01-29
OK, here's a terrible little microbenchmark, which at least demonstrates that it doesn't make a noticeable difference:

Using the exact `scheduler_trace` function that's in the tests for this PR (b675ef3e468600ef66dde43e230ca6f81f75a753), I used the ipython `%%timeit` magic:

	# In [4]: %%timeit
	#   ...: trio.run(scheduler_trace)
	# <with sorting and shuffling enabled>
	7.77 ms ± 1.54 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
	6.79 ms ± 154 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
	6.77 ms ± 119 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
	6.69 ms ± 47.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
	# <without sorting or shuffling>
	7.08 ms ± 1 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
	6.75 ms ± 337 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
	6.46 ms ± 26.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
	7.08 ms ± 1.47 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)

So it's in the noise at that scale.  Now let's try with 1k tasks instead of five:

	# Enabled, 1k instead of five tasks
	808 ms ± 139 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	818 ms ± 134 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	763 ms ± 81 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	777 ms ± 82.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	# Disabled, same scale
	722 ms ± 98.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	666 ms ± 7.34 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	668 ms ± 14.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	670 ms ± 16.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

So we can at least construct workloads where the impact is noticeable (constant work per task and n-log-n scheduling make that easy), though still not pathological.  Judgement time: 

- Is this even a useful benchmark?  Is the "large" scale too large (or too small?), does an IO-free and dependency-free benchmark make sense, etc.  I think this is within a small constant factor of worst-case for a given number of tasks.
- Is a slight slowdown worth the maintenance cost of an internal "could be deterministic" flag and branch?  Probably yes, but I'd appreciate any thoughts.