@njsmith 2019-01-30
So it looks like sorting+shuffling adds ~15% slowdown in a pure-scheduling workload. (Actually I bet we can get that even higher if we switch from `await checkpoint()` to `await cancel_shielded_checkpoint()`, since the latter is more optimized.)

Did you try sorting+shuffling versus shuffling alone, like we currently do?