@Zac-HD 2019-01-30
Shuffling has a slight impact, but more importantly linear rather than log+linear complexity in the number of tasks.  Of course all of this depends on the number of tasks and the actual workload....

	# 1k tasks, shuffling but not sorting
	686 ms ± 6.67 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	698 ms ± 20.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	687 ms ± 12.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
	687 ms ± 12.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Since I expect `pytest-trio` to be the only consumer of determinism here, I've just gone and added a flag to enable the sorting step because it's basically free performance over the other option.