@njsmith 2019-01-30
Yeah, since this is a super-janky API with exactly one consumer, I'm OK with hacky things like this.

I do wonder if it would be better to make the contract *slightly* less janky somehow. Would it be possible to have like `trio._core._run.hypothesis_mode()` that encapsulates both seeding the RNG and turning on sorting? What does the hypothesis RNG seeding functionality actually do, is it just calling `.seed(0)` before each test run or something fancier?