@Zac-HD 2019-01-30
> What does the hypothesis RNG seeding functionality actually do, is it just calling `.seed(0)` before each test run or something fancier?

It *used* to do that, but it's now a bit fancier!  We eventually discovered that repeatedly seeding to zero meant that tests were affecting each other's state, and once we fixed it that there were some rare bugs that just never got triggered when the PRNG was "close" to the zeroed state.  So now we restore the previous state of every PRNG we manage after each test case is executed.

The *other* thing we do is the `random_module()` strategy, which you can use to tell Hypothesis "this test should be affected by randomness".  In that case, we actually draw a random seed value from an internal strategy, so test case executions are varied but still reproducible.  I wrote HypothesisWorks/hypothesis#1741 so we could use this for Trio too without Hypothesis needing to know the details of every such library - we make an exception for Numpy but try to keep everything else generalised.

So I don't think it's worth giving this a nicer API here (at least for now); it's kinda janky but I can add an autouse fixture to pytest-trio that does the monkeypatch for all and only tests with the Trio marker.  Like everything in the Hypothesis, Trio, and Pytest ecosystems it will be beautiful on top and... *pragmatic* underneath - magic hath it's price!