@Zac-HD 2019-02-03
Ping @njsmith; I've fixed the changelog and comment - ready to merge?